requirements.txt
Django~=5.0.3
paho-mqtt~=2.0.0